+++ src/appleseed/foundation/mesh/objmeshfilereader.cpp
@@ -52,7 +52,7 @@ namespace foundation
 
 namespace
 {
-    const size_t Undefined = ~size_t(0);
+    const size_t Undefined = ~0;
 }
 
 struct OBJMeshFileReader::Impl
